URL : /firestation?stationNumber=<firestation_number> refactored
